api/astro_data/requirements.txt
pandas>=1.5.3
python-dotenv>=1.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-huggingface>=0.0.1
langchain-community>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.3.0
scikit-learn>=1.0.0,<1.4.0
joblib>=1.1.0
